requirements.txt
djangorestframework
drf-spectacular
pydantic
firebase-admin
python-dotenv
